scripts/5_finetune_actions.py
#!/usr/bin/env python3
"""
Script 5: Finetune for Continuous Actions

Finetune the foundation model to output continuous robot actions.

Usage:
    python scripts/5_finetune_actions.py \
        foundation_checkpoint=checkpoints/vla_foundation.ckpt \
        experiment=action_finetune
"""

import sys
from pathlib import Path

workspace_root = Path(__file__).parent.parent
sys.path.insert(0, str(workspace_root / "packages"))

import hydra
from omegaconf import DictConfig


@hydra.main(version_base=None, config_path="../config", config_name="config")
def main(cfg: DictConfig):
    """
    Finetune for continuous actions.
    
    TODO: Implement action finetuning:
    1. Load foundation checkpoint
    2. Replace action head
    3. Load robot demonstration dataset
    4. Train with discretized actions
    """
    print("=" * 80)
    print("LAPA Stage 3: Action Finetuning")
    print("=" * 80)
    print("\nTODO: Implement action finetuning (Task 3.4)")
    print("See PLAN.md Section 2.3 for specifications")


if __name__ == "__main__":
    main()

