pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lapa"
version = "0.1.0"
description = "LAPA: Latent Action Pretraining from Videos - A three-stage robot learning system"
readme = "README.md"
requires-python = ">=3.9,<3.12"
license = {text = "MIT"}
authors = [
    {name = "LAPA Team", email = "team@example.com"}
]
keywords = ["robotics", "vision-language-action", "deep-learning", "pytorch"]

dependencies = [
    # Core deep learning
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "pytorch-lightning>=2.1.0",
    "lightning-fabric>=2.1.0",
    
    # Configuration management
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    
    # Data processing
    "webdataset>=0.2.86",
    "pillow>=10.0.0",
    "numpy>=1.24.0,<2.0.0",
    
    # Transformers and vision models
    "transformers>=4.35.0",
    "timm>=0.9.0",
    "accelerate>=0.25.0",
    
    # Experiment tracking and logging
    "wandb>=0.16.0",
    "tensorboard>=2.15.0",
    
    # Utilities
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
    "rich>=13.7.0",
    "einops>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/lapa"
Documentation = "https://lapa.readthedocs.io"
Repository = "https://github.com/your-org/lapa"

[tool.setuptools.packages.find]
where = ["packages"]
include = ["common*", "laq*", "foundation*", "low_level*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=packages --cov-report=html --cov-report=term"

